fix: Support SQS Queue Names (#877)
We run with a mixture of SQS queue names and urls in all environment.

Since we have been passing queue URLs through to the AWS SDK as is, raw queue names lead to errors rather than successful queue operations.

The changes introduced here first check whether the provided queue url or name can be parsed as URL, and if not, tries to retrieve the url for the named queue.

## Related Issues
https://jira.corp.adobe.com/browse/SITES-30735